requirements.txt
langchain
langchain-community
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
# uncomment to use huggingface llms
#huggingface-hub
#transformers
#bitsandbytes
#accelerate

# uncomment to use instructor embeddings
#InstructorEmbedding
sentence-transformers